UBERF-6490: Rework backup tool
+ Do not allow to connect to server until upgrade will be not done.
+ Secure list-workspaces account call
+ Properly handle of server version mismatch
+ Fix wrong value passed to pull request header
