+++ plugins/guest-resources/src/components/GuestApp.svelte
@@ -26,11 +26,11 @@
   {#await connect(getMetadata(workbench.metadata.PlatformTitle) ?? 'Platform')}
     <Loading />
   {:then client}
-    {#if !client && versionError}
+    {#if $versionError}
       <div class="version-wrapper">
         <div class="antiPopup version-popup">
           <h1><Label label={workbench.string.ServerUnderMaintenance} /></h1>
-          {versionError}
+          {$versionError}
         </div>
       </div>
     {:else if client}
