+++ server/backup-service/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "@hcengineering/backup-service",
+  "version": "0.6.0",
+  "main": "lib/index.js",
+  "svelte": "src/index.ts",
+  "types": "types/index.d.ts",
+  "author": "Anticrm Platform Contributors",
+  "license": "EPL-2.0",
+  "scripts": {
+    "build": "compile",
+    "build:watch": "compile",
+    "format": "format src",
+    "run": "ts-node",
+    "test": "jest --passWithNoTests --silent",
+    "_phase:build": "compile transpile src",
+    "_phase:test": "jest --passWithNoTests --silent",
+    "_phase:format": "format src",
+    "_phase:validate": "compile validate"
+  },
+  "devDependencies": {
+    "@hcengineering/platform-rig": "^0.6.0",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-n": "^15.4.0",
+    "eslint": "^8.54.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint-config-standard-with-typescript": "^40.0.0",
+    "prettier": "^3.1.0",
+    "typescript": "^5.3.3",
+    "@types/tar-stream": "^2.2.2",
+    "@types/node": "~20.11.16",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "@types/jest": "^29.5.5"
+  },
+  "dependencies": {
+    "@hcengineering/platform": "^0.6.9",
+    "@hcengineering/core": "^0.6.28",
+    "@hcengineering/contact": "^0.6.20",
+    "@hcengineering/client-resources": "^0.6.23",
+    "@hcengineering/client": "^0.6.14",
+    "@hcengineering/model": "^0.6.7",
+    "tar-stream": "^2.2.0",
+    "@hcengineering/server-tool": "^0.6.0",
+    "@hcengineering/server-core": "^0.6.1",
+    "@hcengineering/server-backup": "^0.6.0",
+    "@hcengineering/minio": "^0.6.0",
+    "@hcengineering/server-token": "^0.6.7"
+  }
+}
